+++ packages/nimble-components/src/drawer/index.ts
@@ -227,7 +227,7 @@ class Drawer extends FoundationDialog {
     }
 }
 
-export const nimbleDrawer = Drawer.compose({
+const nimbleDrawer = Drawer.compose({
     baseName: 'drawer',
     template,
     styles