+++ packages/nimble-components/src/text-field/tests/text-field.spec.ts
@@ -1,0 +1,14 @@
+import {
+    DesignSystem,
+    TextField as FoundationTextField
+} from '@microsoft/fast-foundation';
+import { html } from '@microsoft/fast-element';
+import '../index';
+
+describe('TextField', () => {
+    it('should have its tag returned by tagFor(FoundationTextField)', () => {
+        expect(html`${DesignSystem.tagFor(FoundationTextField)}`.html).toBe(
+            'nimble-text-field'
+        );
+    });
+});